// src/lib/api/model/upsertAiPersonaDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */
import type { UpsertAiPersonaDtoBusinessRules } from "./upsertAiPersonaDtoBusinessRules";

export interface UpsertAiPersonaDto {
  /** Name of the clinic */
  clinicName: string;
  /** Name of the AI agent */
  agentName?: string;
  /** Tone of the AI agent */
  tone?: string;
  /** Message sent when handing off to a human */
  handoffMessage?: string;
  /** Dynamic workflow rules (JSON) */
  businessRules?: UpsertAiPersonaDtoBusinessRules;
}